samples/scripts/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.13.0

# Database
pymongo>=4.3.0
mongoengine>=0.27.0

# Geospatial analysis
shapely>=2.0.0
pyproj>=3.5.0
folium>=0.14.0
geopy>=2.3.0
fiona>=1.9.0  # Required by geopandas for reading shapefiles

# Data fetching and APIs
requests>=2.28.0
aiohttp>=3.8.0

# Data validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# Visualization (for agent testing)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
click>=8.1.0
tqdm>=4.65.0
colorama>=0.4.6

# Testing (optional)
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Linting and formatting (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0 